feat(README.md): add instructions and details about the project
This commit adds a README.md file to the project. The README.md file contains instructions on how to use the project and details about the project. The instructions include installing ngrok and running the honeypot.sh file. The details include information about the two files in the project, the functions defined in the honeypot.sh file, and the function defined in the server.py file. The commit also adds a license file to the project. The license file specifies that the project is licensed under the MIT License.